refactor(billing): move billing orchestration
